fix: macOS Sign-In Methods can't list or remove multiple auth methods
macOS's Sign-In Methods section (net-new in this PR) only ever rendered
the single legacy authType/userAuth fields and had no per-method Remove
button -- the confirmation dialog for it was already wired in
SettingsView.swift but nothing on macOS ever triggered it. A user with
Google+Apple+Seedphrase all linked saw only one method on macOS and
couldn't remove any of them there, while iOS correctly listed and could
remove each one.

Extracted authTypesContains()/parseAuthMethods()/methodDisplayName()
(previously private to SettingsForm-iOS, and partly reimplemented inline
on macOS) into a shared AuthMethods.swift, and rebuilt macOS's Sign-In
Methods section to ForEach over authTypes with a Remove button, matching
iOS.
